requirements.txt
astronomer-cosmos
apache-airflow-providers-snowflake
dbt-snowflake
dbt-utils==1.1.1
yfinance
pandas
numpy
matplotlib
seaborn
plotly
great-expectations
cryptography
keyring
pytest
pytest-cov
flake8
black
isort
mypy
psutil
networkx
pyyaml
